+++ examples/react19/vite.config.js
@@ -18,6 +18,7 @@ export default defineConfig({
     },
   ],
   resolve: {
+    dedupe: ['react', 'react-dom'],
     alias: {
       '@': path.resolve(__dirname, './src'),
       '@mushket-co/block-builder/react': path.resolve(__dirname, '../../src/react.ts'),